fix: headland/center transition
Fix overlapBox() calls, which added a boolean parameter since 22,
so now the switch for exact check is the fourth parameter after the
mask, not the third. This should greatly improve CP 25 in all areas
where the pathfinder is involved.

Add an optional collisionMask parameter to the pathfinder context.
This is now used to retry the pathfinding for the headland/center
transitions with all collisions ignored in case the first try
fails, and thus providing a drivable path to the first waypoint,
even if the vehicle must make a 180 turn.

#79