fix: island boundary vertices
Make sure island boundaries have vertices every few meters,
even if the polygon has long straight edges.

Would also be possible to add waypoints to the generated
course, as this really is a driving problem, not a
generation problem, but the grassfire algorithm generating
the headlands works better with more vertices.