fix: possible specialization race condition
We create a shortcut member (spec_cp<CP specialization name>) for the vehicle
in the specialization's onLoad() function. It seems that sometimes there are
other methods of the specialization are called before onLoad(), primarily when
the vehicle is configured in the shop, so when these methods try to access
the specialization through this shortcut, it is nil.

Introduce a function that returns the specialization, without using the
shortcut and use that to access the specialization everywhere.

#655 #578 #586