fix: refresh markers
The Nexat switches to a different AI direction
node when changing between transport and work
mode. Our markers (and thus, the proximity sensors
are linked to the direction node but this only
happens when the driver starts. Later starts won't
leave the markers alone as they already exist.

Add a refresh marker function which checks if the
direction node of the vehicle changed, and refreshes
the marker nodes when needed. This way, the proximity
sensors are also moved to the current direction node,
although with the Nexat, in work mode they are way too
narrow.