feat: loop turn
Add headland turn maneuver to make corners with a 270 turn.
This is good for rigs that can't reverse but there is plenty of
room on the field to make a 270 loop. Examples are seed drills
with a seed cart. The first headland should be round, the second
and the rest can have a corner and there, this 270 will be used.

Such rigs are currently auto-detected but we can also make it
an option.

#737