perf: pathfinder penalty cache
Once calculated for a grid cell, cache the penalty
in the pathfinder. The penalty depends currently only
on the position and not the heading, so this requires
a limited amount of memory.

This should reduce the calls for getNodePenalty() by
an order of magnitude, but needs to be tested on a
bigger field with difficult situations.