fix: fill level for harvesters
At many harvesters, especially potato and other
root vegetable harvesters, the fruit needs a long
time to get from the ground to the tank.

This means there's a significant amount of fruit
harvested but not showing up in the tank's fill
level.

Add all fruit already harvested to the fill level
so we stop the moment enough fruit is harvested to
reach 100%.

#357