perf: collision avoidance controller
If a harvester has an insanely long course, the
collision avoidance controller, while looking for an
intersection between the unloader and the harvester's
course, iterated until the end of the harvester's
which caused significant slowdowns.

Limit the lookahead for the collision avoidance
controller: always start checking for intersections
at the current waypoint of the unloader and the
harvester, and stop looking after the lookahead
distance.

#871