fix: leaving course after pathfinder track
When CP uses the pathfinder for a connecting track and
there is a headland turn with more than 90 degrees right
after the connecting track, won't move the relevant waypoint
past the corner as the corner waypoint heading points
"backwards" (>90).

Make sure that when resuming from the StartRowOnly turn,
if there is no waypoint in front of the vehicle, use the
one after the turn end wp in the context.

This will still skip the turn maneuver but at least the
driver will continue the course.

#881