fix: check CpUtil.getName parameter
Looks like this function is called with a number in
place of the vehicle during the game load sometime.

Since this is called every time we are logging the
vehicle's name, we have no idea which is the bad call.

Add a check and print the callstack so we can identify
the location. This will also mask the error, so if there
is no other problem, people will not report it, and we'll
never find out...